fix(expert): 修复评审问题
- 后端：ExpertMetadata::resolve_agent_name 统一三处主理 agent 解析，
  修复 team 类型专家在 todo 执行路径不注入的回归（补 unit 测试），
  合并黑板对话路径优先级反了的逻辑。
- 前端：ActionButton 参数输入改为基于 lastGeneratedRef 判断，
  仅在用户未手改 prompt 时实时替换占位符，手动编辑不再被覆盖。
- 前端：ExpertsPanel 1305 行拆分为编排层 + experts/{ExpertCard,
  TeamCard, ExpertDetailModal}；getCategoryName 迁入 types/expert.ts。
- 前端：ExpertBadge 增加 getExpertByNameCached 模块级缓存（含
  请求合并），列表共享同一专家只拉一次；ExpertsPanel 列表刷新
  时统一失效缓存。
- 清理：loop_trigger dispatcher 瘦身为 (runner, db)，消除伪造
  ServiceContext 与空 ExpertIndexManager；handlers/experts.rs
  删除 zip 解压里无效的 canonicalize 死代码；importExpert 返回类型
  修正为 ExpertMetadata | null。
