+++ frontend/src/components/settings/ExpertsPanel.tsx
@@ -1,877 +1,31 @@
 // 专家管理面板：Tab 分区展示专家和专家团，支持卡片点击弹出详情 Modal。
 //
-// 设计思路：
-// - 使用 Ant Design Tabs 分开展示「专家」和「专家团队」，避免混在一起不方便查找
-// - 卡片展示丰富信息：头像、名称、职业、标签、技能数、成员数（团队）等
-// - 点击卡片弹出 Modal 大卡片，展示完整详情（描述、技能、成员列表、标签等）
-// - 充分利用 plugin.json 中的信息：tags、skills、members、categoryId 等
-// - 完整支持亮暗色主题，使用 CSS 变量确保主题适配
+// 卡片（ExpertCard/TeamCard）与详情弹窗（ExpertDetailModal）已拆分到 ./experts/ 下，
+// 本文件只保留列表加载、搜索过滤、导入导出编排与详情 Modal 的状态管理。
 
 import { useState, useEffect, useMemo, useCallback, useRef } from 'react';
-import { Badge, Button, Input, Empty, Spin, Modal, Dropdown, Tabs, Tag, Tooltip, Typography, App } from 'antd';
+import { App, Button, Dropdown, Empty, Input, Modal, Spin, Tabs, Tooltip } from 'antd';
 import type { MenuProps } from 'antd';
 import {
   TeamOutlined,
   UserOutlined,
   ReloadOutlined,
   SearchOutlined,
-  FileTextOutlined,
-  ThunderboltOutlined,
-  DownloadOutlined,
   UploadOutlined,
   FolderOpenOutlined,
-  StarOutlined,
-  RightOutlined,
-  DeleteOutlined,
 } from '@ant-design/icons';
 import { PageCard } from '@/components/common/PageCard';
 import * as db from '@/utils/database';
-import type { ExpertMetadata, SkillMetadata, ExpertMember } from '@/types/expert';
+import type { ExpertMetadata, SkillMetadata } from '@/types/expert';
 import {
   getExpertDisplayName,
   getExpertDescription,
   getExpertProfession,
-  getExpertAvatarUrl,
-  getMemberAvatarUrl,
 } from '@/types/expert';
 import { ExpertCreateModal } from './ExpertCreateModal';
-
-const { Paragraph, Text } = Typography;
-
-/**
- * 获取分类名称映射
- *
- * 根据 plugin.json 中的 categoryId 字段，返回中文分类名称。
- * 分类 ID 格式如 "02-Engineering"、"08-FinanceInvestment"
- */
-function getCategoryName(categoryId?: string): string {
-  if (!categoryId) return '';
-  const map: Record<string, string> = {
-    '02-Engineering': '工程技术',
-    '08-FinanceInvestment': '金融投资',
-  };
-  return map[categoryId] || categoryId.split('-').slice(1).join(' ');
-}
-
-/**
- * 单个专家卡片组件
- *
- * 展示头像、名称、职业、描述、标签、技能数等信息。
- * 卡片悬停时有上浮效果和边框高亮。
- */
-function ExpertCard({ expert, onClick }: {
-  expert: ExpertMetadata;
-  onClick: (expert: ExpertMetadata) => void;
-}) {
-  const displayName = getExpertDisplayName(expert);
-  const profession = getExpertProfession(expert);
-  const description = getExpertDescription(expert);
-  const avatarUrl = getExpertAvatarUrl(expert);
-  const [avatarError, setAvatarError] = useState(false);
-  const showAvatar = avatarUrl && !avatarError;
-
-  return (
-    <div
-      role="button"
-      tabIndex={0}
-      onClick={() => onClick(expert)}
-      onKeyDown={(e) => {
-        if (e.key === 'Enter' || e.key === ' ') {
-          e.preventDefault();
-          onClick(expert);
-        }
-      }}
-      style={{
-        display: 'flex',
-        flexDirection: 'column',
-        gap: 10,
-        padding: 14,
-        borderRadius: 14,
-        border: '1px solid var(--color-border-secondary)',
-        background: 'var(--color-bg-elevated)',
-        cursor: 'pointer',
-        transition: 'all 0.25s cubic-bezier(0.4, 0, 0.2, 1)',
-        height: '100%',
-        overflow: 'hidden',
-      }}
-      onMouseEnter={(e) => {
-        (e.currentTarget as HTMLDivElement).style.transform = 'translateY(-4px)';
-        (e.currentTarget as HTMLDivElement).style.borderColor = 'var(--color-primary)';
-        (e.currentTarget as HTMLDivElement).style.boxShadow = 'var(--shadow-md)';
-      }}
-      onMouseLeave={(e) => {
-        (e.currentTarget as HTMLDivElement).style.transform = 'translateY(0)';
-        (e.currentTarget as HTMLDivElement).style.borderColor = 'var(--color-border-secondary)';
-        (e.currentTarget as HTMLDivElement).style.boxShadow = 'none';
-      }}
-    >
-      {/* 头部：头像 + 名称 + 职业 */}
-      <div style={{ display: 'flex', alignItems: 'flex-start', gap: 10 }}>
-        {showAvatar ? (
-          <img
-            src={avatarUrl}
-            alt={displayName}
-            onError={() => setAvatarError(true)}
-            style={{
-              width: 44,
-              height: 44,
-              borderRadius: 12,
-              objectFit: 'cover',
-              flexShrink: 0,
-              border: '2px solid var(--color-border)',
-            }}
-          />
-        ) : (
-          <div style={{
-            width: 44,
-            height: 44,
-            borderRadius: 12,
-            background: 'linear-gradient(135deg, var(--color-info-bg-1) 0%, var(--color-primary-bg) 100%)',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            flexShrink: 0,
-            border: '2px solid var(--color-border)',
-          }}>
-            <UserOutlined style={{ color: 'var(--color-info)', fontSize: 20 }} />
-          </div>
-        )}
-
-        <div style={{ flex: 1, minWidth: 0 }}>
-          <div style={{ display: 'flex', alignItems: 'center', gap: 6 }}>
-            <Text strong style={{ fontSize: 14, color: 'var(--color-text)' }}>
-              {displayName}
-            </Text>
-            <Tag color="blue" style={{ margin: 0, fontSize: 10, padding: '1px 6px' }}>
-              专家
-            </Tag>
-          </div>
-          {profession && (
-            <div style={{ fontSize: 12, color: 'var(--color-text-secondary)', marginTop: 2 }}>
-              {profession}
-            </div>
-          )}
-        </div>
-      </div>
-
-      {/* 描述：最多 2 行 */}
-      {description && (
-        <div style={{
-          fontSize: 12,
-          color: 'var(--color-text-tertiary)',
-          lineHeight: 1.4,
-          display: '-webkit-box',
-          WebkitLineClamp: 2,
-          WebkitBoxOrient: 'vertical',
-          overflow: 'hidden',
-        }}>
-          {description}
-        </div>
-      )}
-
-      {/* 标签：分类 + 技能标签 */}
-      <div style={{ display: 'flex', flexWrap: 'wrap', gap: 4 }}>
-        {expert.category_id && (
-          <Tag
-            color="purple"
-            style={{ margin: 0, fontSize: 10, padding: '2px 6px', borderRadius: 4 }}
-          >
-            {getCategoryName(expert.category_id)}
-          </Tag>
-        )}
-        {expert.tags && expert.tags.slice(0, 3).map((tag, idx) => (
-          <Tag
-            key={idx}
-            style={{
-              margin: 0,
-              fontSize: 10,
-              padding: '2px 6px',
-              borderRadius: 4,
-              background: 'var(--color-bg-tertiary)',
-              color: 'var(--color-text-secondary)',
-              border: 'none',
-            }}
-          >
-            {tag.zh || tag.en}
-          </Tag>
-        ))}
-      </div>
-
-      {/* 底部：技能数 + 查看详情箭头 */}
-      <div style={{
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'space-between',
-        marginTop: 'auto',
-        paddingTop: 8,
-        borderTop: '1px solid var(--color-border-light)',
-      }}>
-        <div style={{ display: 'flex', alignItems: 'center', gap: 4 }}>
-          <ThunderboltOutlined style={{ fontSize: 12, color: 'var(--color-warning)' }} />
-          <span style={{ fontSize: 12, color: 'var(--color-text-tertiary)' }}>
-            {expert.skills.length} 项技能
-          </span>
-        </div>
-        <RightOutlined style={{ fontSize: 14, color: 'var(--color-text-tertiary)' }} />
-      </div>
-    </div>
-  );
-}
-
-/**
- * 专家团队卡片组件
- *
- * 展示团队头像、名称、负责人、成员数、描述、标签等信息。
- * 团队卡片使用橙色系配色，与单个专家区分。
- */
-function TeamCard({ expert, onClick }: {
-  expert: ExpertMetadata;
-  onClick: (expert: ExpertMetadata) => void;
-}) {
-  const displayName = getExpertDisplayName(expert);
-  const profession = getExpertProfession(expert);
-  const description = getExpertDescription(expert);
-  const avatarUrl = getExpertAvatarUrl(expert);
-  const [avatarError, setAvatarError] = useState(false);
-  const showAvatar = avatarUrl && !avatarError;
-  const memberCount = expert.members?.length || 0;
-  const leadMember = expert.members?.find(m => m.role === 'lead');
-
-  return (
-    <div
-      role="button"
-      tabIndex={0}
-      onClick={() => onClick(expert)}
-      onKeyDown={(e) => {
-        if (e.key === 'Enter' || e.key === ' ') {
-          e.preventDefault();
-          onClick(expert);
-        }
-      }}
-      style={{
-        display: 'flex',
-        flexDirection: 'column',
-        gap: 10,
-        padding: 14,
-        borderRadius: 14,
-        border: '1px solid var(--color-border-secondary)',
-        background: 'var(--color-bg-elevated)',
-        cursor: 'pointer',
-        transition: 'all 0.25s cubic-bezier(0.4, 0, 0.2, 1)',
-        height: '100%',
-        overflow: 'hidden',
-      }}
-      onMouseEnter={(e) => {
-        (e.currentTarget as HTMLDivElement).style.transform = 'translateY(-4px)';
-        (e.currentTarget as HTMLDivElement).style.borderColor = 'var(--color-warning)';
-        (e.currentTarget as HTMLDivElement).style.boxShadow = 'var(--shadow-md)';
-      }}
-      onMouseLeave={(e) => {
-        (e.currentTarget as HTMLDivElement).style.transform = 'translateY(0)';
-        (e.currentTarget as HTMLDivElement).style.borderColor = 'var(--color-border-secondary)';
-        (e.currentTarget as HTMLDivElement).style.boxShadow = 'none';
-      }}
-    >
-      {/* 头部：团队头像 + 名称 + 类型标签 */}
-      <div style={{ display: 'flex', alignItems: 'flex-start', gap: 10 }}>
-        {showAvatar ? (
-          <img
-            src={avatarUrl}
-            alt={displayName}
-            onError={() => setAvatarError(true)}
-            style={{
-              width: 44,
-              height: 44,
-              borderRadius: 12,
-              objectFit: 'cover',
-              flexShrink: 0,
-              border: '2px solid var(--color-border)',
-            }}
-          />
-        ) : (
-          <div style={{
-            width: 44,
-            height: 44,
-            borderRadius: 12,
-            background: 'linear-gradient(135deg, var(--color-warning-bg-1) 0%, var(--color-error-bg) 100%)',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            flexShrink: 0,
-            border: '2px solid var(--color-border)',
-          }}>
-            <TeamOutlined style={{ color: 'var(--color-warning)', fontSize: 20 }} />
-          </div>
-        )}
-
-        <div style={{ flex: 1, minWidth: 0 }}>
-          <div style={{ display: 'flex', alignItems: 'center', gap: 6 }}>
-            <Text strong style={{ fontSize: 14, color: 'var(--color-text)' }}>
-              {displayName}
-            </Text>
-            <Tag color="orange" style={{ margin: 0, fontSize: 10, padding: '1px 6px' }}>
-              团队
-            </Tag>
-          </div>
-          {profession && (
-            <div style={{ fontSize: 12, color: 'var(--color-text-secondary)', marginTop: 2 }}>
-              {profession}
-            </div>
-          )}
-        </div>
-      </div>
-
-      {/* 描述：最多 2 行 */}
-      {description && (
-        <div style={{
-          fontSize: 12,
-          color: 'var(--color-text-tertiary)',
-          lineHeight: 1.4,
-          display: '-webkit-box',
-          WebkitLineClamp: 2,
-          WebkitBoxOrient: 'vertical',
-          overflow: 'hidden',
-        }}>
-          {description}
-        </div>
-      )}
-
-      {/* 团队信息：负责人 + 成员数 */}
-      <div style={{
-        display: 'flex',
-        alignItems: 'center',
-        gap: 8,
-        padding: 8,
-        borderRadius: 8,
-        background: 'var(--color-warning-bg-1)',
-      }}>
-        <div style={{ display: 'flex', alignItems: 'center', gap: 4 }}>
-          <StarOutlined style={{ fontSize: 12, color: 'var(--color-warning)' }} />
-          <span style={{ fontSize: 11, color: 'var(--color-text-secondary)' }}>
-            负责人：{leadMember ? (leadMember.name_zh || leadMember.name_en || '未知') : '无'}
-          </span>
-        </div>
-        <div style={{ width: 1, height: 12, background: 'var(--color-border)' }} />
-        <div style={{ display: 'flex', alignItems: 'center', gap: 4 }}>
-            <UserOutlined style={{ fontSize: 12, color: 'var(--color-warning)' }} />
-            <span style={{ fontSize: 11, color: 'var(--color-text-secondary)' }}>
-              {memberCount} 人团队
-            </span>
-          </div>
-      </div>
-
-      {/* 标签：分类标签 */}
-      {expert.category_id && (
-        <div style={{ display: 'flex', flexWrap: 'wrap', gap: 4 }}>
-          <Tag
-            color="purple"
-            style={{ margin: 0, fontSize: 10, padding: '2px 6px', borderRadius: 4 }}
-          >
-            {getCategoryName(expert.category_id)}
-          </Tag>
-        </div>
-      )}
-
-      {/* 底部：查看详情箭头 */}
-      <div style={{
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'space-between',
-        marginTop: 'auto',
-        paddingTop: 8,
-        borderTop: '1px solid var(--color-border-light)',
-      }}>
-        <div style={{ display: 'flex', alignItems: 'center', gap: 4 }}>
-          <ThunderboltOutlined style={{ fontSize: 12, color: 'var(--color-warning)' }} />
-          <span style={{ fontSize: 12, color: 'var(--color-text-tertiary)' }}>
-            {expert.skills.length} 项技能
-          </span>
-        </div>
-        <RightOutlined style={{ fontSize: 14, color: 'var(--color-text-tertiary)' }} />
-      </div>
-    </div>
-  );
-}
-
-/**
- * 专家详情 Modal 组件
- *
- * 点击卡片后弹出的大卡片，展示完整的专家信息：
- * - 基本信息：头像、名称、职业、描述、版本、分类
- * - 标签列表
- * - 技能列表
- * - 团队成员列表（团队类型）
- * - Agent MD 内容
- *
- * 设计特点：
- * - 使用 Modal 替代 Drawer，更灵活自然
- * - 大卡片布局，信息层次分明
- * - 支持滚动查看长内容
- * - 暗色主题适配完善
- */
-function ExpertDetailModal({
-  open,
-  expert,
-  agentMd,
-  skills,
-  onClose,
-  onExport,
-  exporting,
-  onDelete,
-  deleting,
-}: {
-  open: boolean;
-  expert: ExpertMetadata | null;
-  agentMd: string;
-  skills: SkillMetadata[];
-  onClose: () => void;
-  onExport: () => void;
-  exporting: boolean;
-  onDelete: () => void;
-  deleting: boolean;
-}) {
-  if (!expert) return null;
-
-  const displayName = getExpertDisplayName(expert);
-  const profession = getExpertProfession(expert);
-  const description = getExpertDescription(expert);
-  const avatarUrl = getExpertAvatarUrl(expert);
-  const isTeam = expert.expert_type === 'team';
-  const [avatarError, setAvatarError] = useState(false);
-  const showAvatar = avatarUrl && !avatarError;
-  // 删除确认对话框状态
-  const [deleteConfirmOpen, setDeleteConfirmOpen] = useState(false);
-
-  // 打开删除确认
-  const handleDeleteClick = () => {
-    setDeleteConfirmOpen(true);
-  };
-
-  // 确认删除
-  const handleConfirmDelete = () => {
-    setDeleteConfirmOpen(false);
-    onDelete();
-  };
-
-  return (
-    <>
-      <Modal
-        open={open}
-        onCancel={onClose}
-        footer={null}
-        width={Math.min(720, typeof window !== 'undefined' ? window.innerWidth - 32 : 720)}
-        centered
-        style={{ borderRadius: 'var(--radius-lg)' }}
-      >
-      <div style={{
-        maxHeight: '80vh',
-        overflowY: 'auto',
-        display: 'flex',
-        flexDirection: 'column',
-      }}>
-        {/* 顶部背景区域 */}
-        <div style={{
-          padding: 24,
-          background: isTeam
-            ? 'linear-gradient(135deg, var(--color-warning-bg-1) 0%, var(--color-bg-elevated) 100%)'
-            : 'linear-gradient(135deg, var(--color-info-bg-1) 0%, var(--color-bg-elevated) 100%)',
-          borderBottom: '1px solid var(--color-border-light)',
-        }}>
-          {/* 头部：头像 + 名称 + 操作 */}
-          <div style={{ display: 'flex', alignItems: 'flex-start', justifyContent: 'space-between', gap: 12, flexWrap: 'wrap' }}>
-            <div style={{ display: 'flex', alignItems: 'flex-start', gap: 16, flex: 1, minWidth: 0 }}>
-              {showAvatar ? (
-                <img
-                  src={avatarUrl}
-                  alt={displayName}
-                  onError={() => setAvatarError(true)}
-                  style={{
-                    width: 56,
-                    height: 56,
-                    borderRadius: 14,
-                    objectFit: 'cover',
-                    border: '3px solid var(--color-bg-elevated)',
-                    boxShadow: 'var(--shadow-md)',
-                    flexShrink: 0,
-                  }}
-                />
-              ) : (
-                <div style={{
-                  width: 56,
-                  height: 56,
-                  borderRadius: 14,
-                  background: isTeam
-                    ? 'linear-gradient(135deg, var(--color-warning-bg-1), var(--color-error-bg))'
-                    : 'linear-gradient(135deg, var(--color-info-bg-1), var(--color-primary-bg))',
-                  display: 'flex',
-                  alignItems: 'center',
-                  justifyContent: 'center',
-                  border: '3px solid var(--color-bg-elevated)',
-                  boxShadow: 'var(--shadow-md)',
-                  flexShrink: 0,
-                }}>
-                  {isTeam ? (
-                    <TeamOutlined style={{ color: 'var(--color-warning)', fontSize: 26 }} />
-                  ) : (
-                    <UserOutlined style={{ color: 'var(--color-info)', fontSize: 26 }} />
-                  )}
-                </div>
-              )}
-
-              <div style={{ flex: 1, minWidth: 0 }}>
-                <div style={{ display: 'flex', alignItems: 'center', gap: 8, marginBottom: 6, flexWrap: 'wrap' }}>
-                  <Text strong style={{ fontSize: 18, color: 'var(--color-text)', wordBreak: 'break-word' }}>
-                    {displayName}
-                  </Text>
-                  <Tag color={isTeam ? 'orange' : 'blue'} style={{ fontSize: 11, margin: 0 }}>
-                    {isTeam ? '专家团队' : '单个专家'}
-                  </Tag>
-                  <Tag style={{ fontSize: 11, background: 'var(--color-bg-tertiary)', margin: 0 }}>
-                    v{expert.version}
-                  </Tag>
-                </div>
-                {profession && (
-                  <div style={{ fontSize: 13, color: 'var(--color-text-secondary)', marginBottom: 8 }}>
-                    {profession}
-                  </div>
-                )}
-                {expert.category_id && (
-                  <Tag color="purple" style={{ fontSize: 11 }}>
-                    {getCategoryName(expert.category_id)}
-                  </Tag>
-                )}
-              </div>
-            </div>
-
-            {/* 操作按钮 */}
-            <div style={{ display: 'flex', gap: 8, flexShrink: 0 }}>
-              <Tooltip title="导出为 zip 包">
-                <Button
-                  type="text"
-                  icon={<DownloadOutlined />}
-                  onClick={onExport}
-                  loading={exporting}
-                  style={{ color: 'var(--color-text-secondary)' }}
-                >
-                  导出
-                </Button>
-              </Tooltip>
-              <Tooltip title="删除专家">
-                <Button
-                  type="text"
-                  icon={<DeleteOutlined />}
-                  onClick={handleDeleteClick}
-                  loading={deleting}
-                  style={{ color: 'var(--color-text-secondary)' }}
-                >
-                  删除
-                </Button>
-              </Tooltip>
-            </div>
-          </div>
-
-          {/* 描述 */}
-          {description && (
-            <Paragraph
-              type="secondary"
-              style={{ marginTop: 16, marginBottom: 0, fontSize: 14, lineHeight: 1.6 }}
-            >
-              {description}
-            </Paragraph>
-          )}
-        </div>
-
-        {/* 内容区域 */}
-        <div style={{ padding: 24, display: 'flex', flexDirection: 'column', gap: 20 }}>
-          {/* 标签列表 */}
-          {expert.tags && expert.tags.length > 0 && (
-            <div>
-              <div style={{
-                fontWeight: 600,
-                marginBottom: 10,
-                display: 'flex',
-                alignItems: 'center',
-                gap: 6,
-              }}>
-                <StarOutlined style={{ fontSize: 14, color: 'var(--color-warning)' }} />
-                标签
-              </div>
-              <div style={{ display: 'flex', flexWrap: 'wrap', gap: 6 }}>
-                {expert.tags.map((tag, idx) => (
-                  <Tag
-                    key={idx}
-                    style={{
-                      fontSize: 12,
-                      padding: '4px 10px',
-                      borderRadius: 6,
-                      background: 'var(--color-bg-tertiary)',
-                      border: 'none',
-                    }}
-                  >
-                    {tag.zh || tag.en}
-                  </Tag>
-                ))}
-              </div>
-            </div>
-          )}
-
-          {/* 团队成员列表（仅团队类型） */}
-          {isTeam && expert.members && expert.members.length > 0 && (
-            <div>
-              <div style={{
-                fontWeight: 600,
-                marginBottom: 10,
-                display: 'flex',
-                alignItems: 'center',
-                gap: 6,
-              }}>
-                <UserOutlined style={{ fontSize: 14, color: 'var(--color-warning)' }} />
-                团队成员 ({expert.members.length})
-              </div>
-              <div style={{ display: 'grid', gridTemplateColumns: 'repeat(2, 1fr)', gap: 8 }}>
-                {expert.members.map((member) => (
-                  <MemberItem
-                    key={member.id}
-                    member={member}
-                    isLead={member.role === 'lead'}
-                    expertName={expert.name}
-                  />
-                ))}
-              </div>
-            </div>
-          )}
-
-          {/* 技能列表 */}
-          <div>
-            <div style={{
-              fontWeight: 600,
-              marginBottom: 10,
-              display: 'flex',
-              alignItems: 'center',
-              gap: 6,
-            }}>
-              <ThunderboltOutlined style={{ fontSize: 14, color: 'var(--color-warning)' }} />
-              关联技能 ({skills.length})
-            </div>
-            {skills.length === 0 ? (
-              <Text type="secondary" style={{ fontSize: 13 }}>暂无关联技能</Text>
-            ) : (
-              <div style={{ display: 'flex', flexDirection: 'column', gap: 8 }}>
-                {skills.map((skill) => (
-                  <div
-                    key={skill.skill_name}
-                    style={{
-                      padding: '12px 14px',
-                      borderRadius: 10,
-                      background: 'var(--color-bg-tertiary)',
-                      border: '1px solid var(--color-border-light)',
-                    }}
-                  >
-                    <div style={{ display: 'flex', alignItems: 'center', gap: 8 }}>
-                      <span style={{ fontSize: 16 }}>{skill.yaml_emoji || '⚡'}</span>
-                      <div style={{ flex: 1 }}>
-                        <div style={{ fontWeight: 500, fontSize: 14, color: 'var(--color-text)' }}>
-                          {skill.skill_name}
-                        </div>
-                        {(skill.yaml_description_zh || skill.yaml_description || skill.yaml_description_en) && (
-                          <div style={{
-                            fontSize: 12,
-                            color: 'var(--color-text-secondary)',
-                            marginTop: 4,
-                            lineHeight: 1.5,
-                          }}>
-                            {skill.yaml_description_zh || skill.yaml_description || skill.yaml_description_en}
-                          </div>
-                        )}
-                      </div>
-                    </div>
-                  </div>
-                ))}
-              </div>
-            )}
-          </div>
-
-          {/* Agent MD 内容 */}
-          <div>
-            <div style={{
-              fontWeight: 600,
-              marginBottom: 10,
-              display: 'flex',
-              alignItems: 'center',
-              gap: 6,
-            }}>
-              <FileTextOutlined style={{ fontSize: 14, color: 'var(--color-info)' }} />
-              Agent MD
-            </div>
-            {agentMd ? (
-              <pre style={{
-                background: 'var(--color-bg)',
-                color: 'var(--color-text-secondary)',
-                padding: 14,
-                borderRadius: 10,
-                fontSize: 12,
-                maxHeight: 300,
-                overflow: 'auto',
-                whiteSpace: 'pre-wrap',
-                margin: 0,
-                border: '1px solid var(--color-border)',
-                lineHeight: 1.6,
-              }}>
-                {agentMd}
-              </pre>
-            ) : (
-              <Text type="secondary" style={{ fontSize: 13 }}>暂无 Agent MD 内容</Text>
-            )}
-          </div>
-        </div>
-      </div>
-    </Modal>
-
-    {/* 删除确认对话框 */}
-    <Modal
-      open={deleteConfirmOpen}
-      onCancel={() => setDeleteConfirmOpen(false)}
-      footer={null}
-      width={Math.min(480, typeof window !== 'undefined' ? window.innerWidth - 32 : 480)}
-      centered
-    >
-      <div style={{ padding: '20px 0' }}>
-        <div style={{ display: 'flex', alignItems: 'center', gap: 12, marginBottom: 16 }}>
-          <div style={{
-            width: 48,
-            height: 48,
-            borderRadius: '50%',
-            background: 'var(--color-error-bg-1)',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-          }}>
-            <DeleteOutlined style={{ color: 'var(--color-error)', fontSize: 24 }} />
-          </div>
-          <div>
-            <Text strong style={{ fontSize: 16, color: 'var(--color-text)' }}>
-              确认删除
-            </Text>
-          </div>
-        </div>
-        <Paragraph style={{ fontSize: 14, lineHeight: 1.6, marginBottom: 20 }}>
-          确定要删除「{displayName}」吗？此操作将永久删除该专家及其所有关联数据，<br />
-          <Text strong style={{ color: 'var(--color-error)' }}>删除后无法恢复。</Text>
-        </Paragraph>
-        <div style={{ display: 'flex', gap: 10, justifyContent: 'flex-end' }}>
-          <Button
-            onClick={() => setDeleteConfirmOpen(false)}
-            style={{ minWidth: 80 }}
-          >
-            取消
-          </Button>
-          <Button
-            type="primary"
-            danger
-            onClick={handleConfirmDelete}
-            loading={deleting}
-            style={{ minWidth: 80 }}
-          >
-            确认删除
-          </Button>
-        </div>
-      </div>
-    </Modal>
-    </>
-  );
-}
-
-/**
- * 团队成员项组件
- *
- * 展示单个成员的头像、名称、职业、角色。
- * 负责人使用 Badge.Ribbon 右上角绶带标记，避免在窄屏挤压名称导致布局变形。
- *
- * 头像加载：成员的 avatar_path 是相对路径，需通过后端接口获取二进制。
- * 头像加载失败时回退到默认图标，保证布局不塌陷。
- */
-function MemberItem({
-  member,
-  isLead,
-  expertName,
-}: {
-  member: ExpertMember;
-  isLead: boolean;
-  expertName: string;
-}) {
-  // 头像加载失败状态：true 表示走兜底图标
-  const [avatarError, setAvatarError] = useState(false);
-  // 仅当成员配置了 avatar_path 且未发生加载失败时才尝试展示图片
-  const showAvatar = !!member.avatar_path && !avatarError;
-
-  // 成员卡片内容：头像 + 名称 + 职业
-  const cardContent = (
-    <div style={{
-      display: 'flex',
-      alignItems: 'center',
-      gap: 10,
-      padding: 10,
-      borderRadius: 10,
-      background: isLead ? 'var(--color-warning-bg-1)' : 'var(--color-bg-tertiary)',
-      border: '1px solid var(--color-border-light)',
-      // Ribbon 需要父元素 overflow hidden 才能裁切绶带折角
-      overflow: 'hidden',
-    }}>
-      {showAvatar ? (
-        <img
-          src={getMemberAvatarUrl(expertName, member.id)}
-          alt={member.name_zh || member.name_en || member.id}
-          // 头像加载失败时切换到兜底图标，避免出现破图
-          onError={() => setAvatarError(true)}
-          style={{
-            width: 36,
-            height: 36,
-            borderRadius: '50%',
-            objectFit: 'cover',
-            flexShrink: 0,
-          }}
-        />
-      ) : (
-        <div style={{
-          width: 36,
-          height: 36,
-          borderRadius: '50%',
-          background: isLead ? 'var(--color-warning-bg-1)' : 'var(--color-info-bg-1)',
-          display: 'flex',
-          alignItems: 'center',
-          justifyContent: 'center',
-          flexShrink: 0,
-        }}>
-          {isLead ? (
-            <StarOutlined style={{ color: 'var(--color-warning)', fontSize: 16 }} />
-          ) : (
-            <UserOutlined style={{ color: 'var(--color-info)', fontSize: 16 }} />
-          )}
-        </div>
-      )}
-      <div style={{ flex: 1, minWidth: 0 }}>
-        <Text style={{ fontSize: 13, fontWeight: isLead ? 600 : 500, color: 'var(--color-text)' }}>
-          {member.name_zh || member.name_en || member.id}
-        </Text>
-        {(member.profession_zh || member.profession_en) && (
-          <div style={{ fontSize: 11, color: 'var(--color-text-secondary)', marginTop: 2 }}>
-            {member.profession_zh || member.profession_en}
-          </div>
-        )}
-      </div>
-    </div>
-  );
-
-  // 负责人用 Badge.Ribbon 右上角绶带标记，非负责人直接返回卡片
-  if (isLead) {
-    return (
-      <Badge.Ribbon text="负责人" color="orange">
-        {cardContent}
-      </Badge.Ribbon>
-    );
-  }
-  return cardContent;
-}
+import { ExpertCard } from './experts/ExpertCard';
+import { TeamCard } from './experts/TeamCard';
+import { ExpertDetailModal } from './experts/ExpertDetailModal';
 
 /**
  * 专家管理面板入口
@@ -906,6 +60,9 @@ export function ExpertsPanel() {
 
   // 加载专家列表
   const loadExperts = useCallback(async () => {
+    // 任何触发列表刷新的操作（删除/导入/重载）都可能让单专家缓存 stale，
+    // 这里统一失效 ExpertBadge 的单专家缓存，避免展示过期数据。
+    db.invalidateExpertCache();
     setLoading(true);
     try {
       const data = await db.getAllExperts();
